requirements.txt
django_filter==1.1.0
Django==2.0.2
requests==2.18.1
drf_nested_routers==0.90.2
django_filters==0.2.1
djangorestframework==3.7.7
celery==4.1.0
django-cors-headers==2.1.0

# requirements.txt for HEROKU
gunicorn==19.7.1
whitenoise==3.3.1
honcho==1.0.1
dj-database-url==0.5.0
psycopg2==2.7.4